Add debug logging for lock timeouts
Print additional debug info when we see lock acquisition timing out to
allow users to see what's causing the timeout. We want this to happen in
the process to make sure we have the right view of the filesystem and
can capture the information in the right moment. If we leave this up to
users it becomes tough to debug, as by the time a user remotes into the
machine the conflict is probably already resolved.