Fix behavior of CLI to get versions with status
Right now, if get versions is called with the -s inactive flag Knox it
will still only print active key versions. This is because the backend
assumes the version status to be "Active" by default when "get key" is
called. To get inactive versions, we have to pass the status flag to the
backend otherwise inactive versions will not be visible.